backend/requirements.txt
# ============================================================
# Smart Hasaad Backend - Python Dependencies
# ============================================================

# Core Django
Django==4.2.16
djangorestframework==3.15.2
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Auth
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0

# Async / WebSockets
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
uvicorn[standard]==0.30.6

# Task Queue
celery==5.4.0
redis==5.0.8
django-celery-beat==2.7.0
django-celery-results==2.5.1

# API Documentation
drf-spectacular==0.27.2

# Filtering & Search
django-filter==24.3

# Image Processing
Pillow==10.4.0

# QR Code
qrcode==7.4.2

# Analytics
pandas==2.2.3

# Utilities
python-slugify==8.0.4
shortuuid==1.0.13

# Code Quality
black==24.8.0
isort==5.13.2
flake8==7.1.1

# Testing
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.1
faker==28.0.0

# Production
gunicorn==23.0.0
whitenoise==6.7.0
